Simplify usage (#43)
Simplifies the usages of ScrAPD by removing the subcommands. ScrAPD
wants to follow the Unix philosophy which doing one thing and doing it
well. Therefore, by removing the ability to add sub commands, we ensure
that the focus of this tool will stay on retrieving the data.

The documentation was updating to reflect the changes.