First Name and Nickname parsing (#84)
* use first item of deceased_field for First Name

Changed the part of parse_deceased_field that removed some items and then used the last remaining item as the First Name. I wasn't able to verify that the issue (#74, recording a nickname as the first name) is fixed in the cli.

* make missing name data throw IndexError

* add name with punctuation to parsing test

* move split method within deceased field parser

Allows the parse_deceased_field to be tested by passing in strings, not split lists

* split deceased field on some slashes

To handle gender and ethnicity fields in a format like "W/F"