Parse Age without Date of Birth (#125)
* use first item of deceased_field for First Name

Changed the part of parse_deceased_field that removed some items and then used the last remaining item as the First Name. I wasn't able to verify that the issue (#74, recording a nickname as the first name) is fixed in the cli.

* make missing name data throw IndexError

* add name with punctuation to parsing test

* move split method within deceased field parser

Allows the parse_deceased_field to be tested by passing in strings, not split lists

* split deceased field on some slashes

To handle gender and ethnicity fields in a format like "W/F"

* change parsing tests to expect datetimes

This makes the page content and Twitter parsing function tests expect datetimes.
The goal is to convert date strings to datetimes earlier, so we can more easily compare them chronologically.

* change date string calculations to datetime

* update more tests to expect datetime objects

* add function to parse deceased field without DOB

* run yapf format

* change datetime objects to date objects

None of the datetime objects was using the time fields.

* add tests for formatters

* import date_utils functions by name

* add date serializing to formatters

* run yapf format

* remove ValueError for failed age parsing.

It raises an AttributeError anyway

* slightly delay conversion of datetime to date

This way failure raises the exception expected by the tests

* add line breaks for formatting tests

* rephrase first lines of docstrings for pydocstyle

* use strftime, not isoformat, for date format

* add yapf line break

* delete duplicate test

* change CSV formatter to month-first date format

* change to module-level import of date_utils

* change DOB to datetime.date earlier

* add yapf line breaks