Parametrize retries from the CLI
Adds CLI flags to parametrize the number of attempts and the exponential
backoff multiplier.

Reconfigures the retryer to reraise the last exception instead of
raising `RetryError`, as well as changing the strategy to use a backoff
multiplier. The wait strategy will now be attempt_number * 3 seconds by
default.

The documentation was updated accordingly.

Fixes scrapd/scrapd#149
Fixes scrapd/scrapd#151