Parse notes fields (#166)
* use first item of deceased_field for First Name

Changed the part of parse_deceased_field that removed some items and then used the last remaining item as the First Name. I wasn't able to verify that the issue (#74, recording a nickname as the first name) is fixed in the cli.

* make missing name data throw IndexError

* add name with punctuation to parsing test

* move split method within deceased field parser

Allows the parse_deceased_field to be tested by passing in strings, not split lists

* split deceased field on some slashes

To handle gender and ethnicity fields in a format like "W/F"

* parse notes with beautifulsoup

* add tests for notes fields

fixes #73 and #81

* add bs4 to requirements

* make notes_from_element a separate function

* add equal sign in requirements.txt

* remove check for img tags in notes

These tags should be removed from the notes by BeautifulSoup, and Coveralls says the tests aren't touching this line.

* add tests for parsing to get correct age field

* parse deceased field with BeautifulSoup

Fixes the parsing bug where the Deceased field 'Hispanic male, 19 years of age' was truncated to 'Hispanic male, 19'.

Closes #150.

* add tests for deceased field after parsing

The deceased field wasn't actually in the return value of any function, so in order to test its value, I had to move the line that deletes the deceased field outside the parse_page_content function.

* change import style for bs4

* add test data for stray <strong> tag

* fix variable assignment for parsed deceased field

* convert ValueError to str before storing it

* delete Deceased fields from Twitter descriptions