handle multiple fatalities from collision (#184)
* use first item of deceased_field for First Name

Changed the part of parse_deceased_field that removed some items and then used the last remaining item as the First Name. I wasn't able to verify that the issue (#74, recording a nickname as the first name) is fixed in the cli.

* make missing name data throw IndexError

* add name with punctuation to parsing test

* move split method within deceased field parser

Allows the parse_deceased_field to be tested by passing in strings, not split lists

* split deceased field on some slashes

To handle gender and ethnicity fields in a format like "W/F"

* parse notes with beautifulsoup

* add tests for notes fields

fixes #73 and #81

* add bs4 to requirements

* make notes_from_element a separate function

* add equal sign in requirements.txt

* remove check for img tags in notes

These tags should be removed from the notes by BeautifulSoup, and Coveralls says the tests aren't touching this line.

* add tests for parsing to get correct age field

* parse deceased field with BeautifulSoup

Fixes the parsing bug where the Deceased field 'Hispanic male, 19 years of age' was truncated to 'Hispanic male, 19'.

Closes #150.

* add tests for deceased field after parsing

The deceased field wasn't actually in the return value of any function, so in order to test its value, I had to move the line that deletes the deceased field outside the parse_page_content function.

* change import style for bs4

* add test data for stray <strong> tag

* fix variable assignment for parsed deceased field

* convert ValueError to str before storing it

* delete Deceased fields from Twitter descriptions

* return errors with page content

* split apd to make parsing and regex modules

* don't pass deceased element to parse_page_content

* split person parsing functions into new module

* stop storing lists as values in Deceased dict

* use more specific exceptions in try-except

* defer parsing Deceased field and only do it once

* create parse_person function

This creates a data dict separate from parse_page_content or parse_twitter_description.

* yield multiple decedents' records for one bulletin

* add expected fields for Twitter summary parsing

* start on test for multiple deceased

* save list of Deceased fields during parsing

* recognize multiple deceased if there are two DOBs

Closes #183

* add docstrings

* fix bug from trying to flatten empty list

* remove for loop from twitter description parsing

* split parse_twitter_description into 2 functions

* fix bug where bulletin without DOB failed parsing

* change condition for Notes error message

* delete split_twitter_deceased_field function

* add test for colon in Deceased field

* use regex to remove Deceased label from field text

* make missing data test expect exception

* add unique IDs for fatalities

* fix bug where Deceased field has no <p> tag

* add tests for person parsing errors

* change format at line break

* update README with ID field

* fix bug where record had both "DOB" and "D.O.B."

* yapf format for tests

* refactor splitting field from Twitter summary

to pass flake8